+++ src/setting.py
@@ -1,7 +1,11 @@
 """ Global config """
 DIS_FACTOR = 1/110/1000
 DEBUG_FOLDER = "../debug"
+CACHE_FOLDER = "../cache"
 PANO_FOLFER = '/home/pcl/Data/minio_server/panos'
+
+PRED_MEMO = "/home/pcl/Data/minio_server/input/lane_shape_predict_memo.h5"
+
 
 GBA_BBOX = [112.471628,  22.138605, 114.424664,  23.565487]
 SZ_BBOX  = [113.746280,  22.441466, 114.623972,  22.864722]