+++ src/utils/df_helper.py
@@ -9,7 +9,7 @@ def load_df_memo(fn=None):
     if fn is None:
         return pd.DataFrame()
     
-    df_memo = pd.read_csv(fn)
-    df_memo.loc[:, 'pred'] = df_memo.pred.apply( lambda x: eval(x) )
+    df_memo = pd.read_hdf(fn)
+    # df_memo.loc[:, 'pred'] = df_memo.pred.apply( lambda x: eval(x) )
     
     return df_memo